requirements.txt
fastapi
uvicorn
llama-cpp-python
requests
pydantic
